eliminate otp
